@@ -4,6 +4,11 @@ Tests for AAIP delegation functionality.
 Run with: python -m pytest test_delegation.py
 """
 
+import os
+
+# Import test utilities
+import sys
+
 import pytest
 
 # Import AAIP core modules
@@ -11,17 +16,13 @@ from aaip.core import (
     check_delegation_authorization,
 )
 
-# Import test utilities
-import sys
-import os
-sys.path.insert(0, os.path.join(os.path.dirname(__file__), '..', '..'))
+sys.path.insert(0, os.path.join(os.path.dirname(__file__), "..", ".."))
 from utils import create_test_delegation_payload, create_test_delegation_with_signature
 
 
 class TestDelegationCreation:
     """Test delegation creation and verification."""
-    
-    
+
     def test_delegation_authorization_check(self):
         """Test delegation authorization checking."""
         delegation_payload = create_test_delegation_payload(
@@ -30,30 +31,28 @@ class TestDelegationCreation:
             subject_identity="test_agent",
             subject_identity_system="custom",
             scope=["payments:authorize"],
-            constraints={"max_amount": {"value": 500.0}}
+            constraints={"max_amount": {"value": 500.0}},
         )
-        
+
         # Create a complete delegation with test signature
         complete_delegation = create_test_delegation_with_signature(delegation_payload)
-        
+
         # Test authorization check
-        assert check_delegation_authorization(
-            complete_delegation, 
-            "payments", 
-            "authorize"
-        ) is True
-        
+        assert (
+            check_delegation_authorization(complete_delegation, "payments", "authorize")
+            is True
+        )
+
         # Test with wrong scope
-        assert check_delegation_authorization(
-            complete_delegation, 
-            "email", 
-            "send"
-        ) is False
+        assert (
+            check_delegation_authorization(complete_delegation, "email", "send")
+            is False
+        )
 
 
 class TestDelegationErrorHandling:
     """Test delegation-specific error handling."""
-    
+
     def test_empty_scope(self):
         """Test error handling for empty scope."""
         with pytest.raises(ValueError):
@@ -62,9 +61,9 @@ class TestDelegationErrorHandling:
                 issuer_identity_system="oauth",
                 subject_identity="test_agent",
                 subject_identity_system="custom",
-                scope=[]  # Empty scope should raise ValueError
+                scope=[],  # Empty scope should raise ValueError
             )
-    
+
     def test_invalid_parameters(self):
         """Test error handling for invalid parameters."""
         # Test with empty issuer identity (this should raise ValueError from the internal function)
@@ -74,5 +73,5 @@ class TestDelegationErrorHandling:
                 issuer_identity_system="oauth",
                 subject_identity="test_agent",
                 subject_identity_system="custom",
-                scope=["test"]
+                scope=["test"],
             )
